+++ src/ImageGenerators/PvGenerator.cc
@@ -15,11 +15,10 @@
 
 using namespace carta;
 
-PvGenerator::PvGenerator() : _file_id(0), _name("") {}
+PvGenerator::PvGenerator() : _name("") {}
 
-void PvGenerator::SetFileIdName(int file_id, int index, const std::string& filename, bool is_preview) {
+void PvGenerator::SetFileName(int index, const std::string& filename, bool is_preview) {
     // Optional (not for preview) file id, and name for PV image
-    _file_id = ((file_id + 1) * PV_ID_MULTIPLIER) - index;
     SetPvImageName(filename, index, is_preview);
 }
 
@@ -49,7 +48,7 @@ bool PvGenerator::GetPvImage(const std::shared_ptr<Frame>& frame, const casacore
     image->flush();
 
     // Set returned image
-    pv_image = GeneratedImage(_file_id, _name, image);
+    pv_image = GeneratedImage(_name, image);
     return true;
 }